# Optimize merge tracker and add merge metrics