--- a/packages/beacon-node/src/eth1/eth1MergeBlockTracker.ts
+++ b/packages/beacon-node/src/eth1/eth1MergeBlockTracker.ts
@@ -3,15 +3,24 @@ import {Epoch, RootHex} from "@lodestar/types";
 import {ILogger, isErrorAborted, sleep} from "@lodestar/utils";
 import {toHexString} from "@chainsafe/ssz";
 import {SLOTS_PER_EPOCH} from "@lodestar/params";
+import {IMetrics} from "../metrics/index.js";
 import {ZERO_HASH_HEX} from "../constants/index.js";
+import {TimeSeries} from "../util/timeSeries.js";
 import {IEth1Provider, EthJsonRpcBlockRaw, PowMergeBlock} from "./interface.js";
 import {quantityToNum, quantityToBigint, dataToRootHex} from "./provider/utils.js";
 
 export enum StatusCode {
+  NOT_SET = "NOT_SET",
   PRE_MERGE = "PRE_MERGE",
   SEARCHING = "SEARCHING",
   FOUND = "FOUND",
   POST_MERGE = "POST_MERGE",
+}
+
+export enum MergeBlockCode {
+  NOT_STARTED = "NOT_STARTED",
+  SEARCHING = "SEARCHING",
+  FOUND = "FOUND",
 }
 
 /** Numbers of epochs in advance of merge fork condition to start looking for merge block */
@@ -35,7 +44,13 @@ export type Eth1MergeBlockTrackerModules = {
   signal: AbortSignal;
   clockEpoch: Epoch;
   isMergeTransitionComplete: boolean;
+  metrics: IMetrics | null;
 };
+
+/* maximum time to merge to display = 1 year*/
+const MAX_MERGE_TIME_DISPLAY = 1 * 365 * 24 * 3600;
+/* minimum search range for binary search for merge block*/
+const MIN_SEARCH_RANGE = MAX_BLOCKS_PER_PAST_REQUEST;
 
 /**
  * Follows the eth1 chain to find a (or multiple?) merge blocks that cross the threshold of total terminal difficulty
@@ -44,33 +59,83 @@ export class Eth1MergeBlockTracker {
   private readonly config: IChainConfig;
   private readonly logger: ILogger;
   private readonly signal: AbortSignal;
+  private readonly metrics: IMetrics | null;
+  /** Whether the search is already running, sort of a mutex to avoid eth1 client flooding */
+  private running = false;
 
   /**
    * First found mergeBlock.
    * TODO: Accept multiple, but then handle long backwards searches properly after crossing TTD
    */
   private mergeBlock: PowMergeBlock | null = null;
+  private latestBlock: PowMergeBlock | null = null;
+  private lastSearchedBlock: PowMergeBlock | null = null;
+
   private readonly blocksByHashCache = new Map<RootHex, PowMergeBlock>();
 
   private status: StatusCode = StatusCode.PRE_MERGE;
+  private mergeBlockStatus: MergeBlockCode = MergeBlockCode.NOT_STARTED;
+
   private readonly intervals: NodeJS.Timeout[] = [];
 
+  private mergeSecondsLeft: number | null = null;
+  private mergeTimeseries = new TimeSeries({maxPoints: 100});
+
   constructor(
-    {config, logger, signal, clockEpoch, isMergeTransitionComplete}: Eth1MergeBlockTrackerModules,
+    {config, logger, signal, clockEpoch, isMergeTransitionComplete, metrics}: Eth1MergeBlockTrackerModules,
     private readonly eth1Provider: IEth1Provider
   ) {
     this.config = config;
     this.logger = logger;
     this.signal = signal;
+    this.metrics = metrics;
 
-    // If merge has already happened, disable
+    if (metrics) {
+      const terminalTotalDifficulty = `${this.config.TERMINAL_TOTAL_DIFFICULTY}`;
+      metrics.eth1.eth1MergeStatus.addCollect(() => {
+        // Set merge ttd, merge status and merge block status
+        metrics.eth1.eth1MergeTTD.set(Number(this.config.TERMINAL_TOTAL_DIFFICULTY));
+        metrics.eth1.eth1MergeStatus.set({terminalTotalDifficulty}, Object.keys(StatusCode).lastIndexOf(this.status));
+        metrics.eth1.eth1MergeBlockStatus.set(Object.keys(MergeBlockCode).lastIndexOf(this.mergeBlockStatus));
+
+        if (this.latestBlock !== null) {
+          // Set latestBlock stats
+          metrics.eth1.eth1LatestBlockNumber.set(this.latestBlock.number);
+          metrics.eth1.eth1LatestBlockTD.set(Number(this.latestBlock.totalDifficulty));
+        }
+
+        // Set time to merge
+        if (this.mergeSecondsLeft !== null) {
+          metrics.eth1.eth1TimeLeftForMerge.set(this.mergeSecondsLeft);
+        }
+
+        // Set last searched block metrics
+        if (this.lastSearchedBlock !== null) {
+          metrics.eth1.eth1MergeLastSearchedBlockNumber.set(this.lastSearchedBlock.number);
+          metrics.eth1.eth1MergeLastSearchedBlockTD.set(Number(this.lastSearchedBlock.totalDifficulty));
+        }
+
+        // Set merge block info if found
+        if (this.mergeBlock !== null) {
+          const mergeBlockMeta = {
+            terminalBlockHash: this.mergeBlock.blockHash,
+            // Convert all number/bigints to string labels
+            terminalBlockNumber: `${this.mergeBlock.number}`,
+            terminalBlockTD: `${this.mergeBlock.totalDifficulty}`,
+          };
+          metrics.eth1.eth1MergeBlockDetails.set(mergeBlockMeta, 1);
+        }
+      });
+    }
+
+    // If merge has already happened
     if (isMergeTransitionComplete) {
       this.status = StatusCode.POST_MERGE;
-      return;
     }
 
     // If merge is still not programed, skip
     if (config.BELLATRIX_FORK_EPOCH >= Infinity) {
+      this.status = StatusCode.NOT_SET;
       return;
     }
 
@@ -99,6 +164,15 @@ export class Eth1MergeBlockTracker {
     }
 
     return this.mergeBlock;
+  }
+
+  getMergeTimeLeft(): {mergeSecondsLeft: number; lastUpdate: {time: number; td: number}} | null {
+    if (this.status === StatusCode.NOT_SET || this.mergeSecondsLeft === null) {
+      return null;
+    } else {
+      const [time, td] = this.mergeTimeseries.lastUpdate();
+      return {mergeSecondsLeft: this.mergeSecondsLeft, lastUpdate: {time, td}};
+    }
   }
 
   /**
@@ -139,12 +213,15 @@ export class Eth1MergeBlockTracker {
       totalDifficulty: mergeBlock.totalDifficulty,
     });
     this.mergeBlock = mergeBlock;
-    this.status = StatusCode.FOUND;
+    if (this.status !== StatusCode.POST_MERGE) {
+      this.status = StatusCode.FOUND;
+    }
+    this.mergeBlockStatus = MergeBlockCode.FOUND;
     this.close();
   }
 
   private async startFinding(): Promise<void> {
-    if (this.status !== StatusCode.PRE_MERGE) return;
+    if (this.mergeBlock !== null) return;
 
     // Terminal block hash override takes precedence over terminal total difficulty
     const terminalBlockHash = toHexString(this.config.TERMINAL_BLOCK_HASH);
@@ -163,18 +240,18 @@ export class Eth1MergeBlockTracker {
       return;
     }
 
-    this.status = StatusCode.SEARCHING;
+    if (this.status !== StatusCode.POST_MERGE) {
+      this.status = StatusCode.SEARCHING;
+    }
+    this.mergeBlockStatus = MergeBlockCode.SEARCHING;
+
+    //eth1MergeStatus
     this.logger.info("Starting search for terminal POW block", {
       // eslint-disable-next-line @typescript-eslint/naming-convention
       TERMINAL_TOTAL_DIFFICULTY: this.config.TERMINAL_TOTAL_DIFFICULTY,
     });
 
-    // 1. Fetch current head chain until finding a block with total difficulty less than `transitionStore.terminalTotalDifficulty`
-    this.fetchPreviousBlocks().catch((e) => {
-      if (!isErrorAborted(e)) this.logger.error("Error fetching past POW blocks", {}, e);
-    });
-
-    // 2. Subscribe to eth1 blocks and recursively fetch potential POW blocks
+    // 1. Subscribe to eth1 blocks and find merge block
     const intervalPoll = setInterval(() => {
       this.pollLatestBlock().catch((e) => {
         if (!isErrorAborted(e)) this.logger.error("Error fetching latest POW block", {}, e);
@@ -191,26 +268,41 @@ export class Eth1MergeBlockTracker {
   }
 
   private async fetchPreviousBlocks(): Promise<void> {
-    // If latest block is under TTD, stop. Subscriptions will pick future blocks
-    // If latest block is over TTD, go backwards until finding a merge block
-    // Note: Must ensure parent relationship
-
-    // Fast path for pre-merge scenario
-    const latestBlockRaw = await this.eth1Provider.getBlockByNumber("latest");
-    if (!latestBlockRaw) {
-      throw Error("getBlockByNumber('latest') returned null");
-    }
-
-    const latestBlock = toPowBlock(latestBlockRaw);
     // TTD not reached yet, stop looking at old blocks and expect the subscription to find merge block
-    if (latestBlock.totalDifficulty < this.config.TERMINAL_TOTAL_DIFFICULTY) {
+    if (
+      this.lastSearchedBlock === null ||
+      this.lastSearchedBlock.totalDifficulty < this.config.TERMINAL_TOTAL_DIFFICULTY
+    ) {
       return;
     }
+    this.logger.info(` Finding previous blocks : ${this.lastSearchedBlock.totalDifficulty} `);
 
-    // TTD already reached, search blocks backwards
-    let minFetchedBlockNumber = latestBlock.number;
+    // First do a binary search but lopsided towards the lastSearchedBlock to establish
+    // a new lower bound to lastSearchedBlock where lastSearchedBlock.totalDifficultty
+    // is still >= TERMINAL_TOTAL_DIFFICULTY
+    let searchRange = Math.floor(this.lastSearchedBlock.number * 0.1);
+    while (searchRange > MIN_SEARCH_RANGE && searchRange < this.lastSearchedBlock.number) {
+      const potentialUpdate = this.lastSearchedBlock.number - searchRange;
+      const potentialRawBlock = await this.eth1Provider.getBlockByNumber(potentialUpdate);
+      if (potentialRawBlock === null) {
+        throw Error(`Could not fetch block number=${potentialUpdate}`);
+      }
+      const potentialBlock = toPowBlock(potentialRawBlock);
+      if (potentialBlock.totalDifficulty >= this.config.TERMINAL_TOTAL_DIFFICULTY) {
+        this.lastSearchedBlock = potentialBlock;
+      } else {
+        searchRange = Math.floor(searchRange / 2);
+      }
+    }
+
     // eslint-disable-next-line no-constant-condition
     while (true) {
+      // TTD already reached, search blocks backwards
+      if (this.lastSearchedBlock === null) {
+        throw Error("Internal Error in fetchPreviousBlocks, lastSearchedBlock is null");
+      }
+      const minFetchedBlockNumber: number = this.lastSearchedBlock.number;
+
       const from = Math.max(0, minFetchedBlockNumber - MAX_BLOCKS_PER_PAST_REQUEST);
       // Re-fetch same block to have the full chain of parent-child nodes
       const to = minFetchedBlockNumber;
@@ -223,10 +315,11 @@ export class Eth1MergeBlockTracker {
         if (blocks.length < 2) {
           throw Error(`getBlocksByNumber(${from}, ${to}) returned less than 2 results`);
         }
-
-        for (let i = 0; i < blocks.length - 1; i++) {
+        // Search in reverse so lastSearchedBlock.totalDifficulty >= TTD
+        for (let i = blocks.length - 2; i >= 0; i--) {
           const childBlock = blocks[i + 1];
           const parentBlock = blocks[i];
+          this.lastSearchedBlock = childBlock;
           if (
             childBlock.totalDifficulty >= this.config.TERMINAL_TOTAL_DIFFICULTY &&
             parentBlock.totalDifficulty < this.config.TERMINAL_TOTAL_DIFFICULTY
@@ -243,11 +336,12 @@ export class Eth1MergeBlockTracker {
           }
         }
 
-        // On next round
-        minFetchedBlockNumber = Math.min(to, ...blocks.map((block) => block.number));
-
+        // If we are here it means blocks[0] is still > ttd
+        this.lastSearchedBlock = blocks[0];
         // Scanned the entire blockchain
-        if (minFetchedBlockNumber <= 0) {
+        if (this.lastSearchedBlock.number <= 0) {
+          // Also this is the genesis block, so it needs to be set as the mergeblock
+          this.setTerminalPowBlock(this.lastSearchedBlock);
           return;
         }
       } catch (e) {
@@ -263,13 +357,72 @@ export class Eth1MergeBlockTracker {
    * TODO: How to prevent doing long recursive search after the merge block has happened?
    */
   private async pollLatestBlock(): Promise<void> {
+    if (this.mergeBlock !== null) {
+      // Merge block already found
+      return;
+    }
+
     const latestBlockRaw = await this.eth1Provider.getBlockByNumber("latest");
     if (!latestBlockRaw) {
       throw Error("getBlockByNumber('latest') returned null");
     }
-
     const latestBlock = toPowBlock(latestBlockRaw);
-    await this.fetchPotentialMergeBlock(latestBlock);
+    this.latestBlock = latestBlock;
+
+    // Set time to merge
+    this.mergeTimeseries.addPoint(
+      Number(this.latestBlock.totalDifficulty),
+      quantityToNum(latestBlockRaw.timestamp) * 1000
+    );
+    if (this.latestBlock.totalDifficulty >= this.config.TERMINAL_TOTAL_DIFFICULTY) {
+      this.mergeSecondsLeft = 0;
+    } else if (this.mergeTimeseries.numPoints() > 5) {
+      // Calculate time to merge
+      const mergeDistance = Math.max(
+        Number(this.config.TERMINAL_TOTAL_DIFFICULTY - this.latestBlock.totalDifficulty),
+        0
+      );
+      // Limit the max time by a year, to also avoid division by zero in mergeSecondsLeft calc
+      const tdIncreasePerSecond = Math.max(
+        this.mergeTimeseries.computeLinearSpeed(),
+        (mergeDistance * 1.0) / MAX_MERGE_TIME_DISPLAY
+      );
+      this.mergeSecondsLeft = mergeDistance / tdIncreasePerSecond;
+    }
+
+    // If there is already a search running, return
+    if (this.running) return;
+    this.running = true;
+
+    try {
+      if (latestBlock.totalDifficulty >= this.config.TERMINAL_TOTAL_DIFFICULTY) {
+        // We are post merge, we need to restart search
+        if (this.lastSearchedBlock === null) {
+          this.lastSearchedBlock = latestBlock;
+          await this.fetchPreviousBlocks();
+        } else if (this.lastSearchedBlock.totalDifficulty >= this.config.TERMINAL_TOTAL_DIFFICULTY) {
+          // It seems like the search was previously started but got abandoned because of some
+          // error. Lets just start the search again from this.lastSearchedBlock
+          await this.fetchPreviousBlocks();
+        } else {
+          // Ok, the case is lastSearchedBlock.totalDifficulty < this.config.TERMINAL_TOTAL_DIFFICULTY
+          // We might have got a potential merge block between last searched and the current
+          this.lastSearchedBlock = latestBlock;
+          await this.fetchPotentialMergeBlock(latestBlock);
+        }
+      } else {
+        // No need to run search, latest is behind TTD just update lastSearchedBlock
+        this.lastSearchedBlock = latestBlock;
+      }
+    } catch (e) {
+      this.logger.error(
+        "Search for merge block failed",
+        {latestBlock: latestBlock.number, lastSearchedBlock: this.lastSearchedBlock?.number ?? latestBlock.number},
+        e as Error
+      );
+    }
+
+    this.running = false;
   }
 
   /**
@@ -288,6 +441,7 @@ export class Eth1MergeBlockTracker {
     // Check if this block is already visited
 
     while (block.totalDifficulty >= this.config.TERMINAL_TOTAL_DIFFICULTY) {
+      this.lastSearchedBlock = block;
       if (block.parentHash === ZERO_HASH_HEX) {
         // Allow genesis block to reach TTD
         // https://github.com/ethereum/consensus-specs/pull/2719