<!-- Why is this PR exists? What are the goals of the pull request? -->
Optimize merge tracker and add merge metrics

Closes #4133

### The merge logging, 
curated for brevity: (activated post `bellatrix`) (from local testnet testing ttd=50)
first merge is tracked in %age, and then fully expanded form
expanded form activates at >99.99% merge complete or at `12 hours` before estimated merge (here in this log 10 seconds is the threshold for testing purposes)

```typescript
Eph 0/7 1.000[NODE-A]           info: Synced - slot: 7 - head: 7 0x3794…acb5 - finalized: 0x0000…0000:0 - peers: 0
Eph 1/0 1.000[NODE-A]           info: Synced - slot: 8 - head: 8 0xdf89…ac93 - execution: premerge(td: 64% of 50 - 8 seconds ago) - merge in: ? - finalized: 0x0000…0000:0 - peers: 0
Eph 1/4 1.001[NODE-A]           info: Synced - slot: 12 - head: 12 0x3ef6…c068 - execution: premerge(td: 76% of 50 - 1 seconds ago) - merge in: 30 seconds - finalized: 0x0000…0000:0 - peers: 0
Eph 2/3 1.000[NODE-A]           info: Synced - slot: 19 - head: 19 0xe81c…cb61 - execution: premerge(td: 84% of 50 - 5 seconds ago) - merge in: 20 seconds - finalized: 0x0000…0000:0 - peers: 0
Eph 3/2 1.001[NODE-A]           info: Synced - slot: 26 - head: 26 0x7aa3…5f29 - execution: premerge(td: 48 / 50 - 4 seconds ago) - merge in: 5 seconds - finalized: 0x0000…0000:0 - peers: 0
Eph 4/1 0.235[NODE-A ETH1]      info: Terminal POW block found! hash=0x88a094de38a2c9256def042ae33b6682678b306b8e62949cef520a1733dac79b, number=25, totalDifficulty=50
Eph 4/1 1.001[NODE-A]           info: Synced - slot: 33 - head: 33 0x7006…be28 - execution: premerge(td: 50 / 50 - 13 seconds ago) - merge in: 0 seconds - finalized: 0x95cb…ca69:2 - peers: 0
Eph 4/2 0.053[NODE-A CHAIN]     info: Execution transitioning from PoW to PoS!!!
Eph 4/2 1.004[NODE-A]           info: Synced - slot: 34 - head: 34 0x11c5…a157 - execution: valid(0xb86c…8f4f) - finalized: 0x95cb…ca69:2 - peers: 0
```

### Metrics 

#### Pre-merge 
(local testnet TTD 50)
![image](https://user-images.githubusercontent.com/76567250/181814392-cd178036-4e90-423f-a731-7af56008365c.png)

#### Post-merge 
(local testnet TTD 50)
![image](https://user-images.githubusercontent.com/76567250/181814521-d520f787-47ac-4d1e-8d2a-eda52ee8648c.png)

#### Merge search not started 
(lodestar syncing, not near bellatrix, geth snap syncing fresh)
![image](https://user-images.githubusercontent.com/76567250/181814705-6fbb190d-fda1-48c0-9740-62072532c926.png)

#### Post-merge 
(kiln network)

![image](https://user-images.githubusercontent.com/76567250/181814823-f7f516fb-a6b6-4ad5-9b49-6f189c0fe8f4.png)

#### Merge not set
(mainnet)
![image](https://user-images.githubusercontent.com/76567250/181814967-aa6bd292-d826-4da5-9c25-88139c376b73.png)